requirements.txt
flask
transformers
# torch
torchvision
torchaudio
pillow
opencv-python
numpy
sentencepiece
protobuf